Day 17, star 2.
Work by simulating the x and y movement and capturing "tick ranges" that
define the ticks for which the x and y values are within their
respective components of the target area. If we see the target area for
both x and y values AND their tick ranges overlap, it's a valid vector.